Reverts callback on TypedFunctionType in favor of special case handling for Cretaeobject calls
